siti-slack-bot/requirements.txt
slack-bolt>=1.18.0
anthropic>=0.40.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
google-cloud-bigquery>=3.0.0
